Moving project creation and deletion functions to utils.go
 - Also renames the delete function to make more sense
